requirements.txt
openai
streamlit 
python-dotenv
langchain-openai
langchain_openai
